+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2019-05-15
+- Added XML Documentation validation
+- Added ApiScraper exception to RestrictedFilesValidation
+- Changed outdated documentation
+
 ## [0.4.0] - 2019-04-03
 - Properly handle dependencies on built-in packages, which aren't in the production registry.
 - Fix unit tests